requirements.txt
openai
gradio
pytube
youtube-transcript-api
python-dotenv
langdetect